Reorg types, implementation
1. Creatse top level types.rs for the types, error
   codes common to the crate. Move the error codes, etc
   from utils.rs to the common location
2. Move the protoctol specific trait defines from lib.rs to
   protocol.rs
3. Since the consensus relay implementation has grown big,
   this is moved to separate folder. The types/implementation
   are separated into different files.
